# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Cursos de Finanzas",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuario-roles": {
            "get": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Listar todas las relaciones usuario-rol",
                "operationId": "0de6b0a4bacfb3781d3ff15e26e1e25a",
                "responses": {
                    "200": {
                        "description": "Listado de relaciones"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Asignar un rol a un usuario",
                "operationId": "5b6a884afe7cbdd32480d1ec39edc047",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "id_rol"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer"
                                    },
                                    "id_rol": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relación creada correctamente"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuario-roles/{id}": {
            "put": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Actualizar una relación usuario-rol",
                "operationId": "10054b6b6d39a404775c93b71851b837",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "id_rol"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer"
                                    },
                                    "id_rol": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Relación actualizada correctamente"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UsuarioRol"
                ],
                "summary": "Eliminar una relación usuario-rol",
                "operationId": "e6e657cfe44afcfa96dccdb07b1ca133",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación eliminada correctamente"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token JWT aquí. Ejemplo: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}